SyllabusTimestamp.LastRevision now auto updates.
Removed SyllabusTimestamp insert from CreateCourse() in procedures.sql because it is now handled by a trigger.
Changed most tables with FOREIGN KEYS in tables.sql to CASCADE DELETEs and UPDATEs because I was getting tired of having deleting a Course be a ten minute
process.
Created several TRIGGERs in triggers.sql to catch updates to syllabus
information and update the course's last modified timestamp.
Updated README to reflect new files.